+++ frontend/src/pages/index.tsx
@@ -1,22 +1,43 @@
-import { Table, Tweet } from "../components/table";
+import { ContentCreator } from "../components/contentcreator";
 import React, { useState, useEffect } from "react";
-import { DndProvider } from "react-dnd";
-import { HTML5Backend } from "react-dnd-html5-backend";
-import { Dropzone } from "../components/dropzone";
+import { Dropzone, Tweet } from "../components/dropzone";
 import { InputCard } from "../components/inputCard";
-import { OutputCard } from "../components/outputCard";
+import { OutputCard, OutputEntry, CacheResponse } from "../components/outputCard";
 import { FaKiwiBird } from "react-icons/fa";
-import { OutputEntry } from "../components/itemTypes";
 
 export default function Home() {
   const [droppedTweets, setDroppedTweets] = useState<Tweet[]>([]);
   const [responseData, setResponseData] = useState<OutputEntry | null>(null);
   const [isLoading, setIsLoading] = useState(false);
-  const [apiStatus, setApiStatus] = useState<string>("Offline");
+  const [contexts, setContexts] = useState<{ [key: string]: string }>({});
+  const [prompts, setPrompts] = useState<{ [key: string]: string }>({});
 
-  const handleDrop = (tweet: Tweet) => {
-    setDroppedTweets((prevTweets) => [...prevTweets, tweet]);
-  };
+  const [cache, setCache] = useState<CacheResponse>({});
+
+  useEffect(() => {
+    const fetchData = async () => {
+      const response = await fetch("/api/context");
+      const jsonData = await response.json();
+      setContexts(jsonData);
+    };
+    fetchData();
+  }, []);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      const response = await fetch("/api/prompts");
+      const jsonData = await response.json();
+      setPrompts(jsonData);
+    };
+    fetchData();
+  }, []);
+
+  useEffect(() => {
+    // Initialize the cache from localStorage on component mount
+    const existingCache: CacheResponse = JSON.parse(localStorage.getItem("cache") || "{}");
+    setCache(existingCache);
+
+  }, []);
 
   const handleDeleteTweet = (tweetId: string) => {
     setDroppedTweets((prevTweets) => prevTweets.filter((tweet) => tweet.id !== tweetId));
@@ -29,10 +50,12 @@ export default function Home() {
       input_text: inputText,
       tags: tags,
       prompt: prompt,
+      contexts: contexts,
       tweets: droppedTweets,
     };
 
-    const response = await fetch("http://localhost:8000/process_tweets", {
+    // Note: Change the endpoint to your Next.js API route
+    const response = await fetch("/api/generation", {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
@@ -43,24 +66,22 @@ export default function Home() {
     const data = await response.json();
     setResponseData(data);
     setIsLoading(false);
-  };
 
-  const checkApiHealth = async () => {
     try {
-      const response = await fetch("http://localhost:8000/health");
-      if (response.status === 200) {
-        setApiStatus("Online");
-      } else {
-        setApiStatus("Offline");
-      }
+      const timestamp = new Date().toISOString() + " " + prompt;
+      const existingCache: CacheResponse = JSON.parse(localStorage.getItem("cache") || "{}");
+      existingCache[timestamp] = data;
+
+      localStorage.setItem("cache", JSON.stringify(existingCache));
+
+      // Update the cache state
+      setCache(existingCache);
+
     } catch (error) {
-      setApiStatus("Offline");
+      console.error("Failed to save to cache:", error);
+      // Handle the error appropriately, maybe inform the user, etc.
     }
   };
-
-  useEffect(() => {
-    checkApiHealth();
-  }, []);
 
   const handleAddTweetContent = (content: string) => {
     setDroppedTweets((prevTweets) => [...prevTweets, {
@@ -77,52 +98,47 @@ export default function Home() {
 
   return (
     <div className="min-h-screen">
-      <DndProvider backend={HTML5Backend}>
-        <div className="w-1/3 h-screen bg-zinc-100 fixed shadow-lg overflow-y-auto custom-scrollbar" style={{
-          backgroundImage: "url('/sidebar.png')",
-          backgroundSize: "cover",
-          backgroundPosition: "center",
-        }}>
-          <div className="pt-8 title-container">
-            <h1 className=" text-5xl text-zinc-800 flex items-center justify-center">
-              <FaKiwiBird className="mr-4"></FaKiwiBird>
-              <span className="font-bold text-green-500">
-                Chirp
-              </span>
-              <span className="">Chase</span>
-            </h1>
-            <p className="mt-2 text-sm text-zinc-800 font-mono flex justify-center">
+
+      <div className="w-1/3 h-screen bg-zinc-50 fixed shadow-lg overflow-y-auto custom-scrollbar"
+      >
+        <div className="pt-8 title-container">
+          <h1 className=" text-5xl text-zinc-800 flex items-center justify-center">
+            <FaKiwiBird className="mr-4"></FaKiwiBird>
+            <span className="font-bold text-green-500">
+              Chirp
+            </span>
+            <span className="">Chase</span>
+          </h1>
+          <div className="mt-2 text-xs text-zinc-900 font-mono flex justify-center">
+            <p className="text-sm text-zinc-800 font-mono mr-4 mt-1">
               Analyze and leverage content generation
             </p>
-            <div className="mt-4 text-xs text-zinc-900 font-mono flex justify-center">
-              <span className="rounded-indicator">v0.1.4</span>
-              <span className="rounded-indicator">API: {apiStatus}</span>
-            </div>
-          </div>
-          <div className="pt-6 px-12">
-            <InputCard onSend={handleSend} />
+            <span className="rounded-indicator">v0.2.0</span>
           </div>
         </div>
-        <div
-          className="w-2/3 ml-auto min-h-screen overflow-y-auto custom-scrollbar"
-          style={{
-            backgroundImage: "url('/background.png')",
-            backgroundSize: "cover",
-            backgroundPosition: "center",
-          }}
-        >
-          <div className="container pt-8 px-8 custom-scrollbar">
-            <div className="w-full">
-              <Table onAddTweetContent={handleAddTweetContent} />
-              <Dropzone onDrop={handleDrop} onDelete={handleDeleteTweet} droppedTweets={droppedTweets} />
-              <OutputCard
-                data={responseData}
-                isLoading={isLoading}
-              />
+        <div className="pt-6 px-12">
+          <ContentCreator onAddTweetContent={handleAddTweetContent} />
+          <InputCard onSend={handleSend} context_dict={contexts} prompt_dict={prompts} />
+        </div>
+      </div>
+      <div
+        className="w-2/3 ml-auto min-h-screen dot-grid overflow-y-auto custom-scrollbar bg-zinc-100"
+      >
+        <div className="container pt-8 px-8 custom-scrollbar">
+          <div className="w-full">
+            <div className="min-h-[10vh]">
+
             </div>
+            <Dropzone onDelete={handleDeleteTweet} droppedTweets={droppedTweets} />
+            <OutputCard
+              data={responseData}
+              isLoading={isLoading}
+              cache={cache}
+            />
           </div>
         </div>
-      </DndProvider>
+      </div>
+
     </div>
   );
 }